+++ apps/ledger-live-desktop/src/renderer/screens/exchange/Swap2/App/App.tsx
@@ -5,26 +5,13 @@ import { useLocalLiveAppManifest } from "@ledgerhq/live-common/wallet-api/LocalL
 import React from "react";
 import styled from "styled-components";
 import SwapWebView from "~/renderer/screens/exchange/Swap2/Form/SwapWebViewDemo3";
+import { NetworkErrorScreen } from "~/renderer/components/Web3AppWebview/NetworkError";
+import { useRemoteLiveAppContext } from "@ledgerhq/live-common/platform/providers/RemoteLiveAppProvider/index";
 
 const Root = styled.div`
   display: flex;
   flex-direction: column;
   justify-content: flex-start;
-`;
-
-// TODO: fix with proper error handling
-const ErrorWrapper = styled.div`
-  width: auto;
-  display: inline-flex;
-  align-self: center;
-  align-items: center;
-  justify-self: center;
-  justify-content: center;
-  padding: 24px;
-  border-radius: 14px;
-  background-color: rgba(255, 0, 0, 0.3);
-  color: #fff;
-  font-weight: 500;
 `;
 
 // set the default manifest ID for the production swap live app
@@ -39,11 +26,13 @@ export function SwapApp() {
 
   const localManifest = useLocalLiveAppManifest(swapLiveAppManifestID || undefined);
   const remoteManifest = useRemoteLiveAppManifest(swapLiveAppManifestID || undefined);
+
   const manifest = localManifest || remoteManifest;
 
+  const { updateManifests } = useRemoteLiveAppContext();
+
   if (!manifest) {
-    // TODO: fix with proper error handling
-    return <ErrorWrapper>Unable to load application: missing manifest</ErrorWrapper>;
+    return <NetworkErrorScreen refresh={updateManifests} type="warning" />;
   }
 
   return (