+++ apps/ledger-live-mobile/src/types/common.ts
@@ -2,12 +2,6 @@
 
 import { Unit } from "@ledgerhq/types-cryptoassets";
 
-// FIXME then drop it, i'm not here for that.
-export type T = (
-  arg0: string | null | undefined,
-  arg1: Record<string, unknown> | null | undefined,
-) => string;
-
 export type ConfirmationDefaults = {
   confirmationsNb:
     | {