+++ libs/coin-modules/coin-sui/src/api/index.integration.test.ts
@@ -1,4 +1,9 @@
-import type { AlpacaApi, FeeEstimation, Operation } from "@ledgerhq/coin-framework/api/types";
+import type {
+  AlpacaApi,
+  FeeEstimation,
+  Operation,
+  Pagination,
+} from "@ledgerhq/coin-framework/api/types";
 import { createApi } from ".";
 import { getEnv } from "@ledgerhq/live-env";
 
@@ -34,6 +39,53 @@ describe("Sui Api", () => {
     });
   });
 
+  describe("listOperations, testing cursor logic", () => {
+    // this account has a lot of operations
+    const binance = "0x935029ca5219502a47ac9b69f556ccf6e2198b5e7815cf50f68846f723739cbd";
+
+    async function testListOperations(order: "asc" | "desc" | undefined) {
+      const baseOpts: Pagination = { minHeight: 0 };
+      if (order) {
+        baseOpts.order = order;
+      }
+
+      const [operations1, token1] = await module.listOperations(binance, baseOpts);
+
+      expect(operations1.length).toBeGreaterThan(2);
+      expect(token1).toBeTruthy();
+      const [operations2, _] = await module.listOperations(binance, {
+        ...baseOpts,
+        lastPagingToken: token1,
+      });
+      expect(operations2.length).toBeGreaterThan(2);
+      expect(operations2[0].tx.hash).not.toBe(operations1[0].tx.hash);
+    }
+
+    it("should fetch operations successfully in desc order", async () => {
+      await testListOperations("desc");
+    });
+    it("should fetch operations successfully in asc order", async () => {
+      await testListOperations("asc");
+    });
+    it("should fetch operations successfully in default order", async () => {
+      await testListOperations(undefined);
+    });
+
+    it("shouldn't return cursor on last page", async () => {
+      const [operations, cursor] = await module.listOperations(
+        "0xd8908c165dee785924e7421a0fd0418a19d5daeec395fd505a92a0fd3117e428",
+        { minHeight: 0, order: "asc" },
+      );
+
+      // assume it has not a lot of operations
+      // at time of writing, it has only 2 operations
+      expect(operations.length).toBeLessThan(10);
+      expect(operations.length).toBeGreaterThan(0);
+
+      expect(cursor).toBe("");
+    });
+  });
+
   describe("listOperations", () => {
     let txs: Operation[];
 
@@ -56,11 +108,6 @@ describe("Sui Api", () => {
       expect(checkSet.size).toBeLessThanOrEqual(txs.length);
     });
 
-    it("returns all operations in from the latest order, but sorted in asc", async () => {
-      const [txDesc] = await module.listOperations(SENDER, { minHeight: 0, order: "desc" });
-      expect(txDesc[0]).toStrictEqual(txs[0]);
-    });
-
     it("at least operation should be IN", async () => {
       expect(txs.length).toBeGreaterThanOrEqual(10);
       expect(txs.some(t => t.type === "IN")).toBeTruthy();
@@ -77,6 +124,10 @@ describe("Sui Api", () => {
         order: "asc",
       });
       expect(txs.length).toBeGreaterThanOrEqual(minHeightTxs.length);
+    });
+
+    it("returns block height as a number", async () => {
+      expect(txs.every(t => typeof t.tx.block.height === "number")).toBeTruthy();
     });
   });