+++ libs/ledgerjs/packages/hw-transport-node-speculos-http/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @ledgerhq/hw-transport-node-speculos-http
 
+## 6.30.0
+
+### Minor Changes
+
+- [#11535](https://github.com/LedgerHQ/ledger-live/pull/11535) [`205458e`](https://github.com/LedgerHQ/ledger-live/commit/205458e2d3ca5ec1d7a50d30b63050a692a5e045) Thanks [@sprohaszka-ledger](https://github.com/sprohaszka-ledger)! - Canton add sign hash function
+
+## 6.30.0-next.0
+
+### Minor Changes
+
+- [#11535](https://github.com/LedgerHQ/ledger-live/pull/11535) [`205458e`](https://github.com/LedgerHQ/ledger-live/commit/205458e2d3ca5ec1d7a50d30b63050a692a5e045) Thanks [@sprohaszka-ledger](https://github.com/sprohaszka-ledger)! - Canton add sign hash function
+
 ## 6.29.10
 
 ### Patch Changes